RecyclerView Adapters moved to another package
