+++ app/src/main/java/com/thesis/note/activity/TagEditorActivity.kt
@@ -17,7 +17,7 @@ import kotlinx.android.synthetic.main.template_empty_layout.toolbar
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import com.thesis.note.R
-import com.thesis.note.TagsEditorRecyclerViewAdapter
+import com.thesis.note.recycler_view_adapters.TagsEditorRecyclerViewAdapter
 import com.thesis.note.database.entity.Tag
 import kotlinx.android.synthetic.main.activity_groups_editor.addTagButtonToDb
 import kotlinx.android.synthetic.main.activity_groups_editor.nameOfNewTag
